requirements.txt
fastapi
python-multipart
mammoth
pandas
openpyxl
google-api-python-client
google-auth-oauthlib
python-docx
